+++ public/locales/vn.json
@@ -411,7 +411,16 @@
             "information_settings": "Cài đặt thông tin",
             "organization_name": "Tên tổ chức",
             "phone_number": "Số điện thoại",
-            "organization_address": "Địa chỉ tổ chức"
+            "organization_address": "Địa chỉ tổ chức",
+            "delete_organization": "Xóa tổ chức",
+            "deactivate_organization": "Vô hiệu hóa tổ chức",
+            "organization_deactivation_checkbox": "Xác nhận vô hiệu hóa tổ chức",
+            "organization_deactivation_warning": "Vui lòng xác nhận rằng bạn muốn xóa tổ chức của mình",
+            "organization_deactivation_confirm": "Bạn có chắc chắn muốn hủy đăng ký của mình không?",
+            "organization_deactivation_success": "Đăng ký đã được hủy thành công.",
+            "organization_deactivation_fail": "Đăng ký của bạn đã bị hủy!!",
+            "deleted": "Đã bị xóa",
+            "cancelled": "Đã bị hủy"
         }
     }
 }
