+++ src/views/pages/settings/organization/DeleteOrganizationCard.tsx
@@ -1,0 +1,214 @@
+// ** Next Imports
+import { useRouter } from 'next/router'
+
+// ** React Imports
+import { useState } from 'react'
+
+// ** MUI Imports
+import Box from '@mui/material/Box'
+import Grid from '@mui/material/Grid'
+import Card from '@mui/material/Card'
+import Dialog from '@mui/material/Dialog'
+import Checkbox from '@mui/material/Checkbox'
+import Typography from '@mui/material/Typography'
+import CardHeader from '@mui/material/CardHeader'
+import FormControl from '@mui/material/FormControl'
+import CardContent from '@mui/material/CardContent'
+import DialogContent from '@mui/material/DialogContent'
+import DialogActions from '@mui/material/DialogActions'
+import FormHelperText from '@mui/material/FormHelperText'
+import Button from '@mui/material/Button'
+import FormControlLabel from '@mui/material/FormControlLabel'
+
+// ** Third Party Imports
+import { useForm, Controller } from 'react-hook-form'
+import { useTranslation } from 'react-i18next'
+
+// ** Icon Imports
+import Icon from 'src/@core/components/icon'
+
+// ** Store Imports
+import { useDispatch } from 'react-redux'
+import { AppDispatch } from 'src/store'
+import { deleteOrganization } from 'src/store/apps/organization'
+
+// ** Hook Imports
+import { useCurrentOrganization } from 'src/hooks'
+import { signOut } from 'next-auth/react'
+
+const DeleteOrganizationCard = () => {
+  // ** State
+  const [open, setOpen] = useState<boolean>(false)
+  const [userInput, setUserInput] = useState<string>('yes')
+  const [secondDialogOpen, setSecondDialogOpen] = useState<boolean>(false)
+
+  // ** Hooks
+  const { t } = useTranslation()
+  const router = useRouter()
+  const { organizationId } = useCurrentOrganization()
+  const dispatch = useDispatch<AppDispatch>()
+  const {
+    control,
+    handleSubmit,
+    formState: { errors }
+  } = useForm({ defaultValues: { checkbox: false } })
+
+  const handleLogout = () => {
+    localStorage.removeItem('accessToken')
+    localStorage.removeItem('permissions')
+    signOut({ callbackUrl: '/', redirect: false }).then(() => {
+      router.asPath = '/'
+    })
+  }
+
+  const handleClose = () => setOpen(false)
+
+  const handleSecondDialogClose = () => setSecondDialogOpen(false)
+
+  const onDeleteOrganization = () => setOpen(true)
+
+  const handleConfirmation = (value: string) => {
+    if (value === 'yes') {
+      // Delete organization
+      dispatch(deleteOrganization({ organizationId }))
+      handleLogout()
+    }
+
+    handleClose()
+    setUserInput(value)
+    setSecondDialogOpen(true)
+  }
+
+  return (
+    <>
+      <Grid item xs={12}>
+        <Card>
+          <CardHeader title={t('settings_page.organization.delete_organization')} />
+          <CardContent>
+            <form onSubmit={handleSubmit(onDeleteOrganization)}>
+              <Box sx={{ mb: 4 }}>
+                <FormControl>
+                  <Controller
+                    name='checkbox'
+                    control={control}
+                    rules={{ required: true }}
+                    render={({ field }) => (
+                      <FormControlLabel
+                        label={t('settings_page.organization.organization_deactivation_checkbox')}
+                        sx={errors.checkbox ? { '& .MuiTypography-root': { color: 'error.main' } } : null}
+                        control={
+                          <Checkbox
+                            {...field}
+                            size='small'
+                            name='validation-basic-checkbox'
+                            sx={errors.checkbox ? { color: 'error.main' } : null}
+                          />
+                        }
+                      />
+                    )}
+                  />
+                  {errors.checkbox && (
+                    <FormHelperText sx={{ color: 'error.main' }} id='validation-basic-checkbox'>
+                      {t('settings_page.organization.organization_deactivation_warning')}
+                    </FormHelperText>
+                  )}
+                </FormControl>
+              </Box>
+              <Button variant='contained' color='error' type='submit' disabled={errors.checkbox !== undefined}>
+                {t('settings_page.organization.deactivate_organization')}
+              </Button>
+            </form>
+          </CardContent>
+        </Card>
+      </Grid>
+
+      {/* Deactivate Organization Dialogs */}
+      <Dialog fullWidth maxWidth='xs' open={open} onClose={handleClose}>
+        <DialogContent
+          sx={{
+            pb: theme => `${theme.spacing(6)} !important`,
+            px: theme => [`${theme.spacing(5)} !important`, `${theme.spacing(15)} !important`],
+            pt: theme => [`${theme.spacing(8)} !important`, `${theme.spacing(12.5)} !important`]
+          }}
+        >
+          <Box
+            sx={{
+              display: 'flex',
+              textAlign: 'center',
+              alignItems: 'center',
+              flexDirection: 'column',
+              justifyContent: 'center',
+              '& svg': { mb: 6, color: 'warning.main' }
+            }}
+          >
+            <Icon icon='mdi:alert-circle-outline' fontSize='5.5rem' />
+            <Typography>{t('settings_page.organization.organization_deactivation_confirm')}</Typography>
+          </Box>
+        </DialogContent>
+        <DialogActions
+          sx={{
+            justifyContent: 'center',
+            px: theme => [`${theme.spacing(5)} !important`, `${theme.spacing(15)} !important`],
+            pb: theme => [`${theme.spacing(8)} !important`, `${theme.spacing(12.5)} !important`]
+          }}
+        >
+          <Button variant='contained' sx={{ mr: 2 }} onClick={() => handleConfirmation('yes')}>
+            {t('button.yes')}
+          </Button>
+          <Button variant='outlined' color='secondary' onClick={() => handleConfirmation('cancel')}>
+            {t('button.cancel')}
+          </Button>
+        </DialogActions>
+      </Dialog>
+      <Dialog fullWidth maxWidth='xs' open={secondDialogOpen} onClose={handleSecondDialogClose}>
+        <DialogContent
+          sx={{
+            pb: theme => `${theme.spacing(6)} !important`,
+            px: theme => [`${theme.spacing(5)} !important`, `${theme.spacing(15)} !important`],
+            pt: theme => [`${theme.spacing(8)} !important`, `${theme.spacing(12.5)} !important`]
+          }}
+        >
+          <Box
+            sx={{
+              display: 'flex',
+              alignItems: 'center',
+              flexDirection: 'column',
+              '& svg': {
+                mb: 8,
+                color: userInput === 'yes' ? 'success.main' : 'error.main'
+              }
+            }}
+          >
+            <Icon
+              fontSize='5.5rem'
+              icon={userInput === 'yes' ? 'mdi:check-circle-outline' : 'mdi:close-circle-outline'}
+            />
+            <Typography variant='h4' sx={{ mb: 5 }}>
+              {userInput === 'yes'
+                ? t('settings_page.organization.deleted')
+                : t('settings_page.organization.cancelled')}
+            </Typography>
+            <Typography>
+              {userInput === 'yes'
+                ? t('settings_page.organization.organization_deactivation_success')
+                : t('settings_page.organization.organization_deactivation_fail')}
+            </Typography>
+          </Box>
+        </DialogContent>
+        <DialogActions
+          sx={{
+            justifyContent: 'center',
+            px: theme => [`${theme.spacing(5)} !important`, `${theme.spacing(15)} !important`],
+            pb: theme => [`${theme.spacing(8)} !important`, `${theme.spacing(12.5)} !important`]
+          }}
+        >
+          <Button variant='contained' color='success' onClick={handleSecondDialogClose}>
+            {t('button.ok')}
+          </Button>
+        </DialogActions>
+      </Dialog>
+    </>
+  )
+}
+
+export default DeleteOrganizationCard
